Reset only when asked, only run exec configuration

+++ stuffie/run.sh
@@ -1,8 +1,11 @@
-rm -rf parts
-mkdir parts
+if [[ $2 == 'reset' ]]
+then
+	rm -rf parts
+	mkdir parts
+    echo "Splitting Source file into $1 components..."
+    time split -n $1 --numeric-suffixes "../data/ARC-V1-Feb2018-2/ARC_Corpus.txt" parts/part_
+fi
 
-echo "Splitting Source file into $1 components..."
-time split -n $1 --numeric-suffixes "../data/ARC-V1-Feb2018-2/ARC_Corpus.txt" parts/part_
 
 echo "Compiling.."
 
@@ -11,7 +14,8 @@ cstatus=$?
 echo "Running"
 
 if [ $cstatus -eq 0 ]; then
-    mvn -X exec:java -Dexec.mainClass=it.unibz.inf.stuffie.StuffieConsoleRunner
+    mvn exec:exec
+     # -Dexec.mainClass=it.unibz.inf.stuffie.StuffieConsoleRunner
 else
     echo "Compilation Failed.."
 fi
